+++ src/main/java/com/tsp/new_tsp_admin/api/domain/notice/AdminNoticeDTO.java
@@ -16,8 +16,8 @@ import javax.validation.constraints.NotEmpty;
 @AllArgsConstructor
 @ApiModel(value = "공지사항 관련 변수")
 public class AdminNoticeDTO extends NewCommonDTO {
-    @ApiModelProperty(required = true, value = "rnum", hidden = true, example = "1")
-    private Integer rnum;
+    @ApiModelProperty(required = true, value = "rowNum", hidden = true, example = "1")
+    private Integer rowNum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true, example = "1")
     private Long idx;
